+++ user_service/src/main/java/com/kltn/livability_score/user_service/repository/PreferencePresetRepository.java
@@ -1,6 +1,7 @@
 package com.kltn.livability_score.user_service.repository;
 
 import com.kltn.livability_score.user_service.entity.PreferencePresetEntity;
+import java.math.BigDecimal;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -9,7 +10,9 @@ import org.springframework.stereotype.Repository;
 public interface PreferencePresetRepository extends JpaRepository<PreferencePresetEntity, Long> {
 
   Optional<PreferencePresetEntity> findByPreferenceEducationAndPreferenceSafetyAndPreferenceTransportationAndPreferenceShoppingAndPreferenceEntertainmentAndPreferenceEnvironmentAndPreferenceHealthcare(
-      Float preferenceEducation, Float preferenceSafety, Float preferenceTransportation,
-      Float preferenceShopping, Float preferenceEntertainment, Float preferenceEnvironment,
-      Float preferenceHealthcare);
+      BigDecimal preferenceEducation, BigDecimal preferenceSafety,
+      BigDecimal preferenceTransportation,
+      BigDecimal preferenceShopping, BigDecimal preferenceEntertainment,
+      BigDecimal preferenceEnvironment,
+      BigDecimal preferenceHealthcare);
 }
